backend/requirements.txt
# Sentinel Quant - Backend Dependencies
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Redis & Celery
redis>=5.0.1
celery[redis]>=5.3.4

# Security
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
argon2-cffi>=23.1.0

# Exchange APIs
ccxt>=4.1.0

# HTTP Client
httpx>=0.25.2

# Environment
python-dotenv>=1.0.0

# AI/ML
numpy>=1.26.2
pandas>=2.1.3
torch>=2.1.0
gymnasium>=0.29.1
stable-baselines3>=2.2.1
scikit-learn>=1.3.2
ta>=0.11.0

# Firebase (optional)
# firebase-admin>=6.2.0

# WebSocket
websockets>=12.0